# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API Server for Suasor",
        "title": "Suasor API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Dane Miller",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/admin/client/:clientType/:clientId/test": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Tests the connection to a media client using the provided configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clients"
                ],
                "summary": "Test media client connection",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Client type",
                        "name": "clientType",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Client ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Connection test result",
                        "schema": {
                            "$ref": "#/definitions/responses.APIResponse-responses_TestConnectionResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse-responses_ErrorDetails"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse-responses_ErrorDetails"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse-responses_ErrorDetails"
                        }
                    }
                }
            }
        },
...
        "/clients/media": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves all media client configurations for the user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clients"
                ],
                "summary": "Get all media clients",
                "responses": {
                    "200": {
                        "description": "Media clients retrieved",
                        "schema": {
                            "$ref": "#/definitions/responses.APIResponse-array_responses_ClientResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse-responses_ErrorDetails"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse-responses_ErrorDetails"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new media client configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clients"
                ],
                "summary": "Create a new media client",
                "parameters": [
                    {
                        "description": "Media client data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.SwaggerClientRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Media client created",
                        "schema": {
                            "$ref": "#/definitions/responses.APIResponse-responses_ClientResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse-responses_ErrorDetails"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse-responses_ErrorDetails"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse-responses_ErrorDetails"
                        }
                    }
                }
            }
        },
...
        "/clients/media/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a media client configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clients"
                ],
                "summary": "Delete media client",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Client ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Media client deleted",
                        "schema": {
                            "$ref": "#/definitions/responses.APIResponse-responses_EmptyResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid client ID",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse-responses_ErrorDetails"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse-responses_ErrorDetails"
                        }
                    },
                    "404": {
                        "description": "Client not found",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse-responses_ErrorDetails"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse-responses_ErrorDetails"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "errors.ErrorType": {
            "type": "string",
            "enum": [
                "FAILED_CHECK",
                "UNAUTHORIZED",
                "NOT_FOUND",
                "BAD_REQUEST",
                "INTERNAL_ERROR",
                "FORBIDDEN",
                "CONFLICT",
                "VALIDATION_ERROR",
                "RATE_LIMITED",
                "TIMEOUT",
                "SERVICE_UNAVAILABLE",
                "UNPROCESSABLE_ENTITY"
            ],
            "x-enum-varnames": [
                "ErrorTypeFailedCheck",
                "ErrorTypeUnauthorized",
                "ErrorTypeNotFound",
                "ErrorTypeBadRequest",
                "ErrorTypeInternalError",
                "ErrorTypeForbidden",
                "ErrorTypeConflict",
                "ErrorTypeValidation",
                "ErrorTypeRateLimited",
                "ErrorTypeTimeout",
                "ErrorTypeServiceUnavailable",
                "ErrorTypeUnprocessableEntity"
            ]
        },
...
        "requests.SwaggerClientRequest": {
            "type": "object",
            "required": [
                "clientType",
                "name"
            ],
            "properties": {
                "client": {},
                "clientID": {
                    "type": "integer"
                },
                "clientType": {
                    "$ref": "#/definitions/types.ClientType"
                },
                "isEnabled": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "responses.APIResponse-array_responses_ClientResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responses.ClientResponse"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Operation successful"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "responses.APIResponse-responses_ClientResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/responses.ClientResponse"
                },
                "message": {
                    "type": "string",
                    "example": "Operation successful"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "responses.APIResponse-responses_EmptyResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/responses.EmptyResponse"
                },
                "message": {
                    "type": "string",
                    "example": "Operation successful"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "responses.APIResponse-responses_TestConnectionResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/responses.TestConnectionResponse"
                },
                "message": {
                    "type": "string",
                    "example": "Operation successful"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "responses.ClientResponse": {
            "type": "object",
            "properties": {
                "client": {},
                "clientType": {
                    "$ref": "#/definitions/types.MediaClientType"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
...
        "responses.EmptyResponse": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean"
                }
            }
        },
        "responses.ErrorDetails": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": true
                },
                "stackInfo": {
                    "type": "string"
                }
            }
        },
...
        "responses.ErrorResponse-responses_ErrorDetails": {
            "type": "object",
            "properties": {
                "details": {
                    "$ref": "#/definitions/responses.ErrorDetails"
                },
                "message": {
                    "type": "string",
                    "example": "This is a pretty message"
                },
                "request_id": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer",
                    "example": 201
                },
                "timestamp": {
                    "type": "string"
                },
                "type": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/errors.ErrorType"
                        }
                    ],
                    "example": "FAILED_CHECK"
                }
            }
        },
...
        "responses.TestConnectionResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                },
                "version": {
                    "type": "string"
                }
            }
        },
...
        "types.ClientType": {
            "type": "string",
            "enum": [
                "emby",
                "jellyfin",
                "plex",
                "subsonic",
                "radarr",
                "sonarr",
                "lidarr",
                "unknown",
                "claude",
                "openai",
                "ollama"
            ],
            "x-enum-varnames": [
                "ClientTypeEmby",
                "ClientTypeJellyfin",
                "ClientTypePlex",
                "ClientTypeSubsonic",
                "ClientTypeRadarr",
                "ClientTypeSonarr",
                "ClientTypeLidarr",
                "ClientTypeUnknown",
                "ClientTypeClaude",
                "ClientTypeOpenAI",
                "ClientTypeOllama"
            ]
        },
...
        "types.MediaClientType": {
            "type": "string",
            "enum": [
                "plex",
                "jellyfin",
                "emby",
                "subsonic",
                "unknown"
            ],
            "x-enum-varnames": [
                "MediaClientTypePlex",
                "MediaClientTypeJellyfin",
                "MediaClientTypeEmby",
                "MediaClientTypeSubsonic",
                "MediaClientTypeUnknown"
            ]
        },
...
    }
}